requirements.txt
asn1crypto==0.24.0
click==8.0.4
cryptography==2.1.4
dataclasses==0.8
dlib==19.21.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.3
Flask-Cors==3.0.10
idna==2.6
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==2.0.1
ngt==1.12.2
numpy==1.19.5
Pillow==7.2.0
pybind11==2.10.0
pycrypto==2.6.1
pygobject==3.26.1
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
tqdm==4.64.0
typing-extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0